Bug 1934697 - Replace Raw and Log in navigation pane with GitHub links when available.
generate_commit_info in format.rs already built link to GitHub for a
specific commit.

This re-uses the same data source to build URLs to the GitHub log and
GitHub raw content for the navigation pane.

The Raw link is replaced with the GitHub link when available _instead_
of the Mercurial link.

The Mercurial “Log” link is renamed to “Mercurial Log”.

A “Git Log” link is addded to point to GitHub when available.