Bug 1937445 - increase web-server AMI to 16G, latch mise version
The primary intent here is to boost the size of the web-server AMI to
16G to address running out of disk space during the self-update of the
web-server at the start of its spin-up.

However, when running `./build-docker.sh` locally initially to validate
that the provisioning process was likely to complete on the server, I
encountered a provisioning failure during the installation of `mise`
which is the tool we use to create a sort-of `venv` for node.  mise
seems to be very much under development at this time and we definitely
don't need any of the new enhancements, so I picked the version that
was released on the last successful provisioning run of 2024-Oct-11.

As added to the comment, I'm not sure we really need to be using mise.
I believe my original rationale was that nvm was deprecated and mise
(then rtx) was a rust-based tool that seemed to to support the configs
for the (non-rust) asdf tool and it seemed to work.  Given that we've
updated to Ubuntu 24.04 and our node.js needs are so limited (and the
upstream scip tooling so rarely updated) it might be possible to just
use the default Ubuntu node.js packages.  (I believe I had looked at
getting .deb files for 22.04 and that had eneded up undesirable.)