Nix Flake: Add scip-java
This leverages https://github.com/zaninime/sbt-derivation for building
sbt projects by first building a fixed-output package for the
dependencies then building the actual package.

This is a recurring pattern when packaging software with build systems
that try and download their dependencies themselves.

This also patches the scip-java build scripts to use openjdk from Nix
instead of trying to download it using coursier.