Provide mozsearch tooling via Nix
This adds packages for our Clang plugin and Rust-based tooling and
leverages that on the indexer and web-server.

The Rust parts are built with Crane, which splits the build in two
parts:
- dependencies are built first
- then our own code

This means that unless Cargo.toml or Cargo.lock are touched, or Rust is
upgraded, we don't need to rebuild our Rust dependencies.