Explicitly provide path to wasm_css_analyzer.wasm
Instead of deriving wasmPath from mozSearchRoot in js-analyze.js, this
explicitly passes the path to the directory that contains the
wasm_css_analyzer.wasm binary as an argument to js-analyze.js and as
the MOZSEARCH_WASM_DIR environment variable to js-analyze.sh and
html-analyze.sh.

This will allow packaging wasm_css_analyzer.wasm separately from the
mozsearch scripts.