Nix Flake: Add searchfox index
This one is also peculiar, because it wants to use the history of the
mozsearch repository itself.

When copying a Flake into the store, Nix doesn't copy the .git
directory because Git does not guarantee reproducibility.

Instead, I added a self-with-dotgit non-flake input, which defaults to
/dev/null. If the user leaves it at its default value, the searchfox
index will fail to build with an error along the lines of “failed to
unpack /nix/store/…” when /nix/store/… is empty.

To actually build the searchfox index, the user has to pass
`--override-input self-with-dotgit path:$(pwd)` to instruct Nix to copy
pwd to the store with its .git and use that as-is.