Bug 1959582 - Correct mozsearch-mozilla specific revision checkout for web-server
I'm also speculatively trying to leverage the ability to perform a
shallow checkout of submodules as well to help the disk space usage
for the CI runners.  It sounds like there might have been problems in
very old verisons of git if the revision wasn't the branch head but
that it's no longer a problem.  We also do try and stick to the tip
of the branches, so that also should moot things.