Bug 1890435 - Quiet livegrep protobuf unused struct warnings.
Allowing dead code in the inline-declared module actually quiets them
but I started with seeing if updating the protobuf crate would help,
and that required updating the scip crate since it specifies a protobuf
version with =.  Since we generally want to be up-to-date for both those
crates, I've left that in.

Note that I think the "scip" binary and its print command regressed
after the 0.3.3, but that's a separate issue.