Bug 1890435 - Convert build-blame to use clap for argument parsing.
Support for firefox-main and friends requires that we add some more
command line arguments.  Since this strains the ad-hoc parsing we
were doing and we potentially might want to start dynamically ingesting
blame in the future, it's appropriate to adopt use of clap like we've
done elsewhere to clean this up.

Note that since we already do enable the "env" feature I have
maintained env variable support even though we ideally can end up not
needing to use them in the future.