Bug 1890435 - add a "diagnostics" endpoint to web-server.rs
As part of diagnosing why the new oldrev functionality wasn't working I
thought it would be useful to be able to know what the various revision
map entry counts were, so this adds the ability to retrieve that at
runtime via JSON rather than having to dig it out of logs.

As described in the comments, I think this could also work for an AJAXy
dashboard.  Although the dashboard would be more interested in things
like what revision was indexed, when the indexer started, when the web
server started, things like that which we might actually digest from
the logs at startup or something like that.  That's all future work.