Bug 1989736 - Add an /oldcommit/ endpoint for old m-c /commit/ links
In bug 1890435 I'd punted on this mapping early on in the process
assuming we'd be burning comparatively costly and limited Elastic Load
Balancer rule resources, but we ended up just having all mozilla-central
links sent to the firefox-main tree and having nginx then rewrite the
m-c links to f-m links so /rev/ could become /oldrev/ and unambiguously
be clear about which type of rev we're dealing with.  This means there's
only the additional nginx rule and the marginal increase in web-server
logic, which is mainly down to copy-and-paste.  (I opted to "repeat
yourself" in the interest of clarity.)

When doing the update I noticed that we didn't have path length guards
on hgrev and oldrev to match rev, so I added those.  This changes
what would have been a 500 because of a (caught) panic to a 404.
Note that the rev-flavored guards rely on rust slicing being fine with
slicing at its length to produce an empty slice.

I add /diff/ support in the next commit.

The only other revision-accepting endpoint we have in web-server.rs is
commit-info which is only used for the blame strip hover purposes and
so is only ever same-tree, so we're not adding one for that.  I've added
a comment in the file to that effect since it seems like a reasonable
question someone might have while reading the code.