Bug 1990128 - Make it possible to specify config repo rev when triggering indexer
As conveyed in the comments in the patch, we no longer gracefully handle
when a branch is specified and the branch does not exist in the config
repo.  I've added an optional command-line argument here to allow
mitigating that, but longer term we should try and imitate the original
behavior of falling back to the "master" branch; options in comments.