Bug 1516970 - Historical directory listing.
This implements the /rev endpoint for directories, re-using the same
liquid template as when statically generating them.

The breadcrumb now links to the /rev endpoint when browsing an
historical file version or diff.

I extracted get_object_at into a recursive helper function instead of a
loop. The previous version always returned a String, this nows returns
a (Repositor, Oid) pair that can point to a blob or a tree.
The recursive-ness should also fix browsing nested submodules which I
believe was broken (but didn't actually test).