Build a Git repo for the tests index sources
This is required to test some of Searchfox's Git-related features, such
as browsing historical file and directory versions, Git blame support
and soon reworked Git-based code coverage.

Unfortunately the Git SHA leaks into test expectations. This is
mitigated by having a known author/committer and dates, but the SHA
will change everytime changes are made in tests/tests/files.

test_SearchBoxLayout is updated because we now have a revision in the
file view of the tests repo, but not in the folder view.